add htlc subcommads and split system contract to different files
